add functionality and documentation

+++ remove.py
@@ -1,6 +1,8 @@
 from sys import argv
 from shutil import rmtree
 from list import get_node_versions
+from urllib import request, error
+import json
 from os import path
 from consts import github_repo_link
 
@@ -9,6 +11,15 @@ def remove_node_version():
         print("Version argument is also needed.")
         exit()  
     version_number = argv[2]
+    if version_number == 'latest':
+        url = request.urlopen(f"https://nodejs.org/download/release/index.json")
+        data = json.load(url)
+        version_number = data[0]['version'][1:]
+    if version_number == 'lts':
+        url = request.urlopen(f"https://nodejs.org/download/release/index.json")
+        data = json.load(url)
+        filtered_version = filter(lambda version: version['lts'] != False, data)
+        version_number = next(filtered_version)['version'][1:]
     if (get_node_versions().count(version_number) != 0):
         try:
             rmtree(path.expanduser(f"~/node/node-v{version_number}-linux-x64"))
